Allineato con xFactory\OldVersions\Longiano\XFactoryNET alla versione DevExpress 20.2

+++ XFactoryNET.Custom.MM1.WcfService/SvcPesa.svc.cs
@@ -12,7 +12,7 @@ namespace XFactoryNET.Custom.MM1.WcfService
     public class SvcPesa : XFactoryNET.ServiceLibrary.ISvcPesa
     {
 
-        public float LeggiPeso(string codicePesa)
+        public decimal LeggiPeso(string codicePesa)
         {
             return ThreadPesa.GetPeso(codicePesa);
         }
